sec 2: mention some future work for scan
The comment asked more to explain why the current implementation discards all but the last element in the first phase. There isn't a particular reason for that one way or the other, the implementation simply has not been optimised. This comment just points that optimisation out as useful future work.