Update description of the Ray trace program
Include a not about the very high register usage and corresponding low thread occupancy, which is probably the main source of the poor performance.

Remove the sentence about evaluating both sides of the branch, as that has been fixed (in what commit? was this already fixed in the previous benchmarks, or are the submodules not linked correctly??)