Update dotp numbers with an additional test
This was a manual hack of the generated code to make sure that 32-bit counters were used in the body of the code. The new code updated the original in the cache, and confirmed that the new version was used.

It turns out that this had no impart on performance whatsoever, even though the generated PTX did not contain the conversion instructions from 64 to 32 bits. Maybe on this device those instructions are NOPs.